+++ src/constants.ts
@@ -1,4 +1,4 @@
-export const VERSION = "4.1.0";
+export const VERSION = "4.1.1";
 
 export const Status = {
   QUEUED: "queued",